Move Windows only classes to their own namespace
