+++ CommandLine/src/GameRunner/Windows/ISteamService.cs
@@ -1,4 +1,4 @@
-﻿namespace Worms.GameRunner
+﻿namespace worms.GameRunner.Windows
 {
     public interface ISteamService
     {
