+++ CommandLine/src/GameRunner/Windows/SteamService.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading;
 
-namespace Worms.GameRunner
+namespace worms.GameRunner.Windows
 {
     internal class SteamService : ISteamService
     {
