+++ CommandLine/src/GameRunner/Windows/WindowTools.cs
@@ -6,7 +6,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace Worms.GameRunner
+namespace worms.GameRunner.Windows
 {
     public static class WindowTools
     {
