+++ CommandLine/src/GameRunner/Windows/WormsLocator.cs
@@ -2,8 +2,9 @@
 using System.Diagnostics;
 using System.IO;
 using Microsoft.Win32;
+using Worms.GameRunner;
 
-namespace Worms.GameRunner
+namespace worms.GameRunner.Windows
 {
     internal class WormsLocator : IWormsLocator
     {
